Move testServiceToService function
