+++ test/servinge2e/service_to_service_test.go
@@ -1,7 +1,11 @@
 package servinge2e
 
 import (
+	"context"
 	"testing"
+
+	meta "k8s.io/apimachinery/pkg/apis/meta/v1"
+	pkgTest "knative.dev/pkg/test"
 
 	"github.com/openshift-knative/serverless-operator/test"
 	network "knative.dev/networking/pkg"
@@ -55,5 +59,93 @@ func TestServiceToServiceCalls(t *testing.T) {
 		t.Run(scenario.name, func(t *testing.T) {
 			testServiceToService(t, caCtx, testNamespace, scenario)
 		})
+	}
+}
+
+func testServiceToService(t *testing.T, ctx *test.Context, namespace string, tc testCase) {
+	// Create a ksvc with the specified annotations and labels
+	service := test.Service(tc.name, namespace, helloworldImage, tc.annotations)
+	service.ObjectMeta.Labels = tc.labels
+	service, err := ctx.Clients.Serving.ServingV1().Services(namespace).Create(context.Background(), service, meta.CreateOptions{})
+	if err != nil {
+		t.Errorf("error creating ksvc: %v", err)
+		return
+	}
+
+	// Let the ksvc be deleted after test
+	ctx.AddToCleanup(func() error {
+		t.Logf("Cleaning up Knative Service '%s/%s'", service.Namespace, service.Name)
+		return ctx.Clients.Serving.ServingV1().Services(namespace).Delete(context.Background(), service.Name, meta.DeleteOptions{})
+	})
+
+	// Wait until the Ksvc is ready.
+	service, err = test.WaitForServiceState(ctx, service.Name, service.Namespace, test.IsServiceReady)
+	if err != nil {
+		t.Errorf("error waiting for ksvc readiness: %v", err)
+		return
+	}
+
+	serviceURL := service.Status.URL.URL()
+
+	// For cluster-local ksvc, we deploy an "HTTP proxy" service, and request that one instead
+	if service.GetLabels()[network.VisibilityLabelKey] == string(nv1alpha1.IngressVisibilityClusterLocal) {
+		// Deploy an "HTTP proxy" towards the ksvc (using an httpproxy image from knative-serving testsuite)
+		httpProxy, err := ctx.Clients.Serving.ServingV1().Services(namespace).Create(context.Background(),
+			httpProxyService(tc.name+"-proxy", namespace, service.Status.URL.Host), meta.CreateOptions{})
+		if err != nil {
+			t.Errorf("error creating ksvc: %v", err)
+			return
+		}
+
+		// Let the ksvc be deleted after test
+		ctx.AddToCleanup(func() error {
+			t.Logf("Cleaning up Knative Service '%s/%s'", httpProxy.Namespace, httpProxy.Name)
+			return ctx.Clients.Serving.ServingV1().Services(namespace).Delete(context.Background(), httpProxy.Name, meta.DeleteOptions{})
+		})
+
+		// Wait until the Proxy is ready.
+		httpProxy, err = test.WaitForServiceState(ctx, httpProxy.Name, httpProxy.Namespace, test.IsServiceReady)
+		if err != nil {
+			t.Errorf("error waiting for ksvc readiness: %v", err)
+			return
+		}
+
+		serviceURL = httpProxy.Status.URL.URL()
+	}
+
+	// Verify the service is actually accessible from the outside
+	if _, err := pkgTest.WaitForEndpointState(
+		context.Background(),
+		&pkgTest.KubeClient{Kube: ctx.Clients.Kube},
+		t.Logf,
+		serviceURL,
+		pkgTest.EventuallyMatchesBody(helloworldText),
+		"WaitForRouteToServeText",
+		true); err != nil {
+		t.Errorf("the Route at domain %s didn't serve the expected text %q: %v", service.Status.URL.URL(), helloworldText, err)
+	}
+
+	// Verify the expected istio-proxy is really there
+	podList, err := ctx.Clients.Kube.CoreV1().Pods(namespace).List(context.Background(), meta.ListOptions{LabelSelector: "serving.knative.dev/service=" + service.Name})
+	if err != nil {
+		t.Errorf("error listing pods: %v", err)
+		return
+	}
+
+	for _, pod := range podList.Items {
+		istioProxyFound := false
+		for _, container := range pod.Spec.Containers {
+			if container.Name == "istio-proxy" {
+				istioProxyFound = true
+			}
+		}
+
+		if tc.expectIstioSidecar != istioProxyFound {
+			if tc.expectIstioSidecar {
+				t.Errorf("TestCase %s expects istio-proxy to be present, but no such container exists in %s", tc.name, pod.Name)
+			} else {
+				t.Errorf("TestCase %s does not expect istio-proxy to be present in pod %s, but it has one", tc.name, pod.Name)
+			}
+		}
 	}
 }
